Extract validation decorator internals
